messing with service worker / firebase code
I think the notifications send twice because of the onBackgroundMessage() function collecting/displaying notification but FCM protocol already does this...

https://github.com/firebase/quickstart-js/blob/master/messaging/index.html
https://github.com/nuxt-community/firebase-module/issues/536

"These fields map to keys available through the Message object. For example, key/value pairs entered in the Custom data field of the composer are handled as a data payload for the notification. Other fields map directly to keys in the notification object or in the platform specific notification configuration"
-firebase api https://firebase.google.com/docs/cloud-messaging/js/send-with-console